+++ gtsam/navigation/NavStateImuEKF.cpp
@@ -17,7 +17,8 @@
  * @authors Derek Benham, Frank Dellaert
  */
 
-#include <gtsam/navigation/NavStateImuEKF.h> 
+#include <gtsam/base/Matrix.h>
+#include <gtsam/navigation/NavStateImuEKF.h>
 
 namespace gtsam {
 
@@ -41,27 +42,51 @@ Vector9 navStateImuDynamics(const NavState& X, const Vector3& gyro,
 
   if (H) {
     H->setZero();
-    // Helper: skew-symmetric matrix
-    auto skew = [](const Vector3& a) {
-      Matrix3 S;
-      S << 0.0, -a.z(), a.y(), a.z(), 0.0, -a.x(), -a.y(), a.x(), 0.0;
-      return S;
-    };
-
     // xi_rot = gyro -> no dependence on state
-    // Rows 0:3 already zero
-
     // xi_trans = R^T v
     // d(xi_trans)/d(dR) = +skew(R^T v) ; d(xi_trans)/d(dV) = I ; d/d(dP) = 0
-    H->template block<3, 3>(3, 0) = skew(v_body);
-    H->template block<3, 3>(3, 6) = Matrix3::Identity();
-
+    H->template block<3, 3>(3, 0) = skewSymmetric(v_body);
+    H->template block<3, 3>(3, 6) = I_3x3;
     // xi_vel = a + R^T g
     // d(xi_vel)/d(dR) = +skew(R^T g)
-    H->template block<3, 3>(6, 0) = skew(g_body);
+    H->template block<3, 3>(6, 0) = skewSymmetric(g_body);
   }
 
   return xi;
+}
+
+// ===== NavStateImuEKF methods =====
+
+NavStateImuEKF::NavStateImuEKF(
+    const NavState& X0, const Covariance& P0,
+    const std::shared_ptr<PreintegrationParams>& params)
+    : Base(X0, P0), params_(params) {
+  // Build process noise Q_ = block_diag(Cg, Ci, Ca)
+  const Matrix3& Cg = params_->gyroscopeCovariance;
+  const Matrix3& Ci = params_->integrationCovariance;
+  const Matrix3& Ca = params_->accelerometerCovariance;
+  Q_.setZero();
+  Q_.template block<3, 3>(0, 0) = Cg;
+  Q_.template block<3, 3>(3, 3) = Ci;
+  Q_.template block<3, 3>(6, 6) = Ca;
+}
+
+void NavStateImuEKF::predict(const Vector3& gyro, const Vector3& accel,
+                             double dt) {
+  auto dyn = [&](const NavState& X, OptionalJacobian<Dim, Dim> H) {
+    return navStateImuDynamics(X, gyro, accel, params_->n_gravity, H);
+  };
+  Base::predict(dyn, dt, Q_);
+}
+
+const std::shared_ptr<PreintegrationParams>& NavStateImuEKF::params() const {
+  return params_;
+}
+
+const Vector3& NavStateImuEKF::gravity() const { return params_->n_gravity; }
+
+const NavStateImuEKF::Covariance& NavStateImuEKF::processNoise() const {
+  return Q_;
 }
 
 }  // namespace gtsam
